requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
passlib[bcrypt]
python-jose[cryptography]
pydantic
email-validator
python-multipart
python-dotenv
